+++ src/main/java/com/data/dataxer/models/dto/BankAccountDTO.java
@@ -3,12 +3,14 @@ package com.data.dataxer.models.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 public class BankAccountDTO {
     private Long id;
     Integer bankCode;
-    Integer accountNumber;
+    BigInteger accountNumber;
     String bankName;
     String currency;
     String iban;
